fix(perf-obs): address review comments from github-advanced-security
- Remove wave/config/changelog.json from git index (gitignored generated file)
- render_perf_alloy_config: validate api_key_env_var is an identifier to
  satisfy CodeQL clear-text-storage check; add discovery.relabel blocks
  for wave and perf_exporter scrape targets so run-identity labels apply
  to all scrape jobs, not just the unix exporter
- perf.yml: install Alloy into $RUNNER_TEMP/bin instead of /usr/local/bin
  to avoid requiring sudo; add continue-on-error: true to upsert step
- upsert_grafana_dashboard: remove unused `import copy`; wrap urlopen in
  try/except for URLError/HTTPError so provisioning failures do not fail CI
- perf_metrics_exporter: rename wave_perf_requests_total gauge to
  wave_perf_requests_count to comply with Prometheus naming conventions
  (_total suffix implies a counter, not a gauge)
- MetricsPrometheusServletJakartaIT: move JVM/process binder assertions
  from duplicate Test class into IT; convert Test to a lightweight stub
